+++ src/x12sdk/v5010/x12_834_005010X220A1/parsing.py
@@ -369,7 +369,7 @@ def set_provider_information_loop(
         return
 
     coverage = _get_coverage(context)
-    if TransactionLoops.MEMBER_COVERAGE_PROVIDER_INFORMATION not in context:
+    if TransactionLoops.MEMBER_COVERAGE_PROVIDER_INFORMATION not in coverage:
         coverage[TransactionLoops.MEMBER_COVERAGE_PROVIDER_INFORMATION] = []
 
     coverage[TransactionLoops.MEMBER_COVERAGE_PROVIDER_INFORMATION].append(